requirements.txt
# Load .env file
python-dotenv

# FastAPI
fastapi
uvicorn
jinja2
pydantic

# Generative AI
torch
tiktoken
langchain
langchain-openai
langchain-chroma
langchain-faiss